refactor: generator modules
