+++ generator/saofile.js
@@ -1,85 +1,71 @@
 const path = require('path')
-const glob = require('glob')
 const validate = require('validate-npm-package-name')
 const configs = require('../template.config')
-const {mergeJson, sortObj} = require('./utils')
-
-const getDockerRunScript = () => {
-  const scripts = {
-    postinstall: 'sh ./.postinstall.sh',
-  }
-
-  return {
-    scripts,
-  }
-}
+const {getModules} = require('./modules')
 
 module.exports = {
   prompts() {
     if (this.sao.opts.mock) return []
     const {config} = this.sao.opts
-    return [
-      ...(config.folder
-        ? []
-        : [
-            {
-              name: 'folder',
-              type: 'input',
-              message: 'Your project name',
-            },
-          ]),
-      ...(config.template
-        ? []
-        : [
-            {
-              name: 'template',
-              type: 'list',
-              choices: configs.map(c => ({
-                name: c.template,
-                value: c.template,
-              })),
-              message: 'Choose a template',
-            },
-          ]),
-      ...('language' in config
-        ? []
-        : [
-            {
-              name: 'language',
-              type: 'list',
-              choices: ['JavaScript', 'TypeScript'].map(v => ({
-                name: v,
-                value: v,
-              })),
-              message: 'Programming language',
-            },
-          ]),
-      ...('docker' in config
-        ? []
-        : [
-            {
-              name: 'docker',
-              type: 'confirm',
-              default: true,
-              message: 'Dockerize or not',
-            },
-          ]),
-      ...('cypress' in config
-        ? []
-        : [
-            {
-              name: 'cypress',
-              type: 'confirm',
-              default: true,
-              message: 'Use cypress or not',
-            },
-          ]),
+
+    const prompts = [
+      {
+        skip: () => config.folder,
+        prompt: {
+          name: 'folder',
+          type: 'input',
+          message: 'Your project name',
+        },
+      },
+      {
+        skip: () => config.template,
+        prompt: {
+          name: 'template',
+          type: 'list',
+          choices: configs.map(c => ({
+            name: c.template,
+            value: c.template,
+          })),
+          message: 'Choose a template',
+        },
+      },
+      {
+        skip: () => 'language' in config,
+        prompt: {
+          name: 'language',
+          type: 'list',
+          choices: ['JavaScript', 'TypeScript'].map(v => ({
+            name: v,
+            value: v,
+          })),
+          message: 'Programming language',
+        },
+      },
+      {
+        skip: () => 'docker' in config,
+        prompt: {
+          name: 'docker',
+          type: 'confirm',
+          default: true,
+          message: 'Dockerize or not',
+        },
+      },
+      {
+        skip: () => 'cypress' in config,
+        prompt: {
+          name: 'cypress',
+          type: 'confirm',
+          default: true,
+          message: 'Use cypress or not',
+        },
+      },
     ]
+
+    return prompts.filter(p => !p.skip()).map(p => p.prompt)
   },
   templateData() {
     return this.sao.opts.config
   },
-  // TODO: 这个方法太大了，需要抽离子方法
   actions() {
     const {
       sao: {opts},
@@ -114,190 +100,7 @@ module.exports = {
     }
     opts.outDir = path.resolve(opts.outDir, opts.config.folder)
 
-    const resolveDir = dir => path.resolve(__dirname, '../template', dir)
-
-    const addBaseFramework = {
-      type: 'add',
-      files: '**',
-      templateDir: resolveDir('framework-base'),
-    }
-    // 这些配置文件在模板中是加了'_'前缀的（防止影响到本项目），移到生成的项目后要去掉前缀
-    // 同时也防止这些文件无法发布至 npm
-    const restoreConfigsName = {
-      type: 'move',
-      patterns: {
-        '_.eslintrc.js': '.eslintrc.js',
-        '_.gitignore': '.gitignore',
-        '_package.json': 'package.json',
-      },
-    }
-
-    const addCustomFramework = {
-      type: 'add',
-      files: '**',
-      templateDir: resolveDir(`framework-${opts.config.template}`),
-    }
-    // 合并基础模板和特定模板的 package.json
-    const mergePackageJson = [
-      {
-        type: 'modify',
-        files: 'package.json',
-        handler(basePackage) {
-          const customPackage = require(path.resolve(
-            opts.outDir,
-            '_package.json',
-          ))
-          let result = mergeJson(basePackage, {
-            ...customPackage,
-            name: opts.config.folder,
-          })
-          if (opts.config.docker) {
-            result = mergeJson(result, getDockerRunScript())
-          }
-          result['create-nuxt-app'] = require('../package.json').version
-          ;['dependencies', 'devDependencies'].forEach(k => sortObj(result[k]))
-          return result
-        },
-      },
-      {
-        type: 'remove',
-        files: '_package.json',
-      },
-    ]
-
-    const addModules = Object.keys(opts.config)
-      .filter(k => !['folder', 'template'].includes(k))
-      .map(k => ({
-        type: 'add',
-        files: '**',
-        templateDir: resolveDir(`modules/${opts.config[k]}`),
-      }))
-    // 生成的nuxt项目的srcDir配置=src。这里我们把除test以外的文件夹移动到src目录下
-    const moveDirsToSrc = {
-      type: 'move',
-      patterns: {
-        ...[addBaseFramework, addCustomFramework, ...addModules]
-          .filter(a => a.templateDir)
-          .map(a =>
-            glob.sync('!(test)/', {
-              cwd: a.templateDir,
-            }),
-          )
-          .reduce((dirs1, dirs2) => [...dirs1, ...dirs2])
-          .reduce((res, dir) => ({...res, [dir]: `src/${dir}`}), {}),
-      },
-    }
-
-    const moveJestEslintrc = {
-      type: 'move',
-      patterns: {
-        [path.resolve(
-          opts.outDir,
-          'test/unit/_.eslintrc.js',
-        )]: 'test/unit/.eslintrc.js',
-        '_.babelrc': '.babelrc',
-      },
-    }
-
-    let addDockerFile = []
-    if (opts.config.docker) {
-      addDockerFile = [
-        {
-          type: 'add',
-          files: '**',
-          templateDir: resolveDir('docker-cmd'),
-        },
-      ]
-    }
-    let addE2eModule = []
-    if (opts.config.cypress) {
-      const cypressModulePath = resolveDir('modules/cypress')
-
-      addE2eModule = [
-        {
-          type: 'add',
-          files: '**',
-          templateDir: resolveDir('modules/cypress'),
-        },
-        {
-          type: 'move',
-          patterns: {
-            [path.resolve(
-              opts.outDir,
-              'test/e2e/_.eslintrc.js',
-            )]: 'test/e2e/.eslintrc.js',
-          },
-        },
-        {
-          type: 'modify',
-          files: 'package.json',
-          handler(basePackage) {
-            const customPackage = require(path.resolve(
-              cypressModulePath,
-              '_package.json',
-            ))
-
-            let result = mergeJson(basePackage, customPackage)
-
-            ;['dependencies', 'devDependencies'].forEach(k =>
-              sortObj(result[k]),
-            )
-            return result
-          },
-        },
-        {
-          type: 'remove',
-          files: '_package.json',
-        },
-      ]
-    }
-
-    let addTypeScriptModule = []
-    if (opts.config.language === 'TypeScript') {
-      const typescriptModulePath = resolveDir('modules/typescript')
-
-      addTypeScriptModule = [
-        {
-          type: 'add',
-          files: '**',
-          templateDir: resolveDir('modules/typescript'),
-        },
-        {
-          type: 'modify',
-          files: 'package.json',
-          handler(basePackage) {
-            const customPackage = require(path.resolve(
-              typescriptModulePath,
-              '_package.json',
-            ))
-
-            let result = mergeJson(basePackage, customPackage)
-
-            ;['dependencies', 'devDependencies'].forEach(k =>
-              sortObj(result[k]),
-            )
-            return result
-          },
-        },
-        {
-          type: 'remove',
-          files: ['_package.json', 'jsconfig.json'],
-        },
-      ]
-    }
-
-    return [
-      addBaseFramework,
-      restoreConfigsName,
-      addCustomFramework,
-      mergePackageJson,
-      addModules,
-      moveDirsToSrc,
-      moveJestEslintrc,
-      addDockerFile,
-      addE2eModule,
-      addTypeScriptModule,
-    ].reduce((r, a) => r.concat(a), []) // 和 flat（node >= 11.15.0) 效果一样，性能差点
+    return getModules(opts)
   },
   completed() {
     const cd = () => {
